+++ Lavender/FurnitureLib/FurnitureCreator.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using Lavender.RuntimeImporter;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -70,36 +71,121 @@ namespace Lavender.FurnitureLib
                 _resArea[i] = (Furniture.BuildingArea)furnitureData.restrictedAreas[i];
             }
 
-            // Get the AssetBundle and loads the needed Assets
-            var fileStream = new FileStream(furnitureData.assetBundlePath, FileMode.Open, FileAccess.Read);
-            var assetBundle = AssetBundle.LoadFromStream(fileStream);
-            if (assetBundle == null)
-            {
-                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get AssetBundle at '{furnitureData.assetBundlePath}'!");
-                return null;
-            }
+            GameObject prefab;
+            GameObject previewPrefab ;
+            Sprite image;
 
-            Sprite image = assetBundle.LoadAsset<Sprite>(furnitureData.imageName);
-            GameObject prefab = assetBundle.LoadAsset<GameObject>(furnitureData.prefabName);
-            GameObject previewPrefab = assetBundle.LoadAsset<GameObject>(furnitureData.previewPrefabName);
+            if(furnitureData.assetBundlePath.EndsWith(".json"))
+            {
+                if(File.Exists(furnitureData.assetBundlePath))
+                {
+                    try
+                    {
+                        string rawFurnitureAssetData = File.ReadAllText(furnitureData.assetBundlePath);
+                        FurnitureAssetData? furnitureAssetData = JsonConvert.DeserializeObject<FurnitureAssetData>(rawFurnitureAssetData);
+                        if (furnitureAssetData != null)
+                        {
+                            // get the directory path of the json file
+                            string path = furnitureData.assetBundlePath.Substring(0, furnitureData.assetBundlePath.Length - Path.GetFileName(furnitureData.assetBundlePath).Length);
 
-            if (image == null)
-            {
-                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get image '{furnitureData.imageName}'!");
-                return null;
-            }
-            if (prefab == null)
-            {
-                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get prefab '{furnitureData.prefabName}'!");
-                return null;
-            }
-            if (previewPrefab == null)
-            {
-                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get preview prefab '{furnitureData.previewPrefabName}'!");
-                return null;
-            }
+                            // Load Sprite
+                            Sprite? s = ImageLoader.LoadSprite(path + furnitureAssetData.imagePath);
+                            if (s != null) image = s;
+                            else
+                            {
+                                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get image '{path + furnitureAssetData.imagePath}'!");
+                                return null;
+                            }
 
-            assetBundle.Unload(false);
+                            // Load the OBJ and create the needed prefabs
+                            GameObject obj = new GameObject($"{furnitureData.title} - OBJ");
+
+                            MeshRenderer meshRenderer = obj.AddComponent<MeshRenderer>();
+                            MeshFilter meshFilter = obj.AddComponent<MeshFilter>();
+
+                            try
+                            {
+                                if (!File.Exists(path + furnitureAssetData.objPath)) return null;
+
+                                Mesh mesh = FastObjImporter.Instance.ImportFile(path + furnitureAssetData.objPath);
+                                mesh.name = furnitureData.title + "_Mesh";
+                                meshFilter.mesh = mesh;
+
+                                List<Material> materials = new List<Material>();
+                                foreach(string texturePath in furnitureAssetData.texturePaths)
+                                {
+                                    Material mat = new Material(Shader.Find("Standard"));
+                                    mat.mainTexture = ImageLoader.LoadImage(path + texturePath);
+
+                                    materials.Add(mat);
+                                }
+
+                                meshRenderer.materials = materials.ToArray();
+                            }
+                            catch(Exception e)
+                            {
+                                LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': {e}");
+                                return null;
+                            }
+
+                            prefab = GameObject.Instantiate(obj);
+                            prefab.AddComponent<MeshCollider>();
+
+                            previewPrefab = GameObject.Instantiate(obj);
+                            var boxCollider = previewPrefab.AddComponent<BoxCollider>();
+                            boxCollider.isTrigger = true;
+                            var rb = previewPrefab.AddComponent<Rigidbody>();
+                            rb.isKinematic = true;
+                        }
+                        else
+                        {
+                            LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': Error while deserializing FurnitureAssetData at '{furnitureData.assetBundlePath}'!");
+                            return null;
+                        }
+                    }
+                    catch(Exception e)
+                    {
+                        LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': {e}");
+                        return null;
+                    }
+                }
+                else
+                {
+                    LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get FurnitureAssetData at '{furnitureData.assetBundlePath}'!");
+                    return null;
+                }
+            }
+            else
+            {
+                // Get the AssetBundle and loads the needed Assets
+                var fileStream = new FileStream(furnitureData.assetBundlePath, FileMode.Open, FileAccess.Read);
+                var assetBundle = AssetBundle.LoadFromStream(fileStream);
+                if (assetBundle == null)
+                {
+                    LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get AssetBundle at '{furnitureData.assetBundlePath}'!");
+                    return null;
+                }
+
+                image = assetBundle.LoadAsset<Sprite>(furnitureData.imageName);
+                prefab = assetBundle.LoadAsset<GameObject>(furnitureData.prefabName);
+                previewPrefab = assetBundle.LoadAsset<GameObject>(furnitureData.previewPrefabName);
+
+                if (image == null)
+                {
+                    LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get image '{furnitureData.imageName}'!");
+                    return null;
+                }
+                if (prefab == null)
+                {
+                    LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get prefab '{furnitureData.prefabName}'!");
+                    return null;
+                }
+                if (previewPrefab == null)
+                {
+                    LavenderLog.Error($"Error while creating furniture '{furnitureData.title}': couldn't get preview prefab '{furnitureData.previewPrefabName}'!");
+                    return null;
+                }
+            }
 
             // Creates the Furniture
             Furniture furniture = NewFurniture(
@@ -142,6 +228,10 @@ namespace Lavender.FurnitureLib
                         f.addressableAssetPath = $"Lavender<#>{json_path}";
 
                         return f;
+                    }
+                    else
+                    {
+                        LavenderLog.Error($"FurnitureCreator.Create(): Error while deserializing FurnitureConfig at '{json_path}'!");
                     }
                 }
                 catch (Exception e)